server/requirements.txt
django==1.10
djangorestframework
django-cors-middleware
